magit-log-select-arguments: add --graph to default value
And stop hard-coding `--graph' in `magit-commit-squash-internal'.
This makes it possible for users NOT to use `--graph' if that has
negative effects on performance.  On the other hand it also means
that all uses of `magit-log-select' now show the graph by default,
but the latter might actually be beneficial too.

Re #3559.